refactor: refactor components structure
