+++ src/components/Card/PersonInfo.tsx
