+++ src/components/Header/Header.tsx
@@ -3,7 +3,7 @@ import './Header.scss';
 import React from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
 
-import { THeaderProps } from '../types/component';
+import { THeaderProps } from '../../types/component';
 
 enum Title {
   HOME = 'Home',
