+++ src/layouts/RootLayout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 
-import HeaderWithTitle from '../components/Header';
+import HeaderWithTitle from '../components/Header/Header';
 
 class RootLayout extends React.Component {
   render() {
