+++ src/pages/Home.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Cards from '../components/cards/Cards';
-import SearchBar from '../components/SearchBar';
+import Cards from '../components/Cards/Cards';
+import SearchBar from '../components/SearchBar/SearchBar';
 
 class Home extends React.Component {
   render() {
